[PY] Change dtype to int64 in flag generators

+++ gph/python/ripser_interface.py
@@ -484,16 +484,16 @@ def ripser_parallel(X, maxdim=1, thresh=np.inf, coeff=2, metric="euclidean",
 
     if ret_representative_simplices:
         finite_0 = np.array(res.flag_persistence_generators_by_dim.finite_0,
-                            dtype=np.int32).reshape(-1, 3)
+                            dtype=np.int64).reshape(-1, 3)
         finite_higher = [
-            np.array(x, dtype=np.int32).reshape(-1, 4)
+            np.array(x, dtype=np.int64).reshape(-1, 4)
             for x in res.flag_persistence_generators_by_dim.finite_higher
             ]
         essential_0 = \
             np.array(res.flag_persistence_generators_by_dim.essential_0,
-                     dtype=np.int32)
+                     dtype=np.int64)
         essential_higher = [
-            np.array(x, dtype=np.int32).reshape(-1, 2)
+            np.array(x, dtype=np.int64).reshape(-1, 2)
             for x in res.flag_persistence_generators_by_dim.essential_higher
             ]
         ret['rpsm'] = (finite_0, finite_higher, essential_0, essential_higher)
